feat: Delete buttons

+++ eGuide.AdminUI/src/app/components/service-components/service-list/service-list.component.ts
@@ -4,6 +4,7 @@ import { Store } from '@ngrx/store';
 import { ToastrService } from 'ngx-toastr';
 import { Service } from 'src/app/models/service';
 import { setServiceEditData } from 'src/app/state/service-edit-data/service-edit-data.action';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-service-list',
@@ -30,14 +31,31 @@ export class ServiceListComponent implements OnInit {
   }
 
   deleteService(id: string) {
-    this.serviceService.deleteService(id).subscribe({
-      next: () => {
-        this.getService();
-        this.toastr.success('Service deleted successfully');
-      },
-      error: () => {
-        this.toastr.error('Service not deleted');
-      },
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.serviceService.deleteService(id).subscribe({
+          next: () => {
+            this.getService();
+            this.toastr.success('Service deleted successfully');
+            Swal.fire({
+              title: 'Deleted!',
+              text: 'Your file has been deleted.',
+              icon: 'success',
+            });
+          },
+          error: (err) => {
+            this.toastr.error(err.error);
+          },
+        });
+      }
     });
   }
